backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
boto3==1.18.54
botocore==1.21.54
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
django-hosts==4.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-sort-order-field==1.2
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.20.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.4.0
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.1.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-monkey-business==1.0.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
watchtower==1.0.6
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
